Migrate sed/phot compute functions from load_lc_mock.py to new module sed_from_mock.py. This commit does not break any tests in pytest, but it does break the demo_diffsky_recompute_from_mock.ipynb doctest by introducing an ImportError.
